fix: Update test packages to comply with testpackage linter
- Change package declarations from `hal` to `hal_test` for external API tests
- Change package declarations from `store` to `store_test` for store tests
- Move tests accessing internal functions/fields to `*_internal_test.go` files:
  - Move TestNewSunTimes (accesses sunTimes.location) to sun_internal_test.go
  - Move getShortFunctionName and getStringOrStringSlice tests to util_internal_test.go
- Update imports and function calls to use proper package prefixes
- Remove util_test.go (tests moved to util_internal_test.go)
- Add hal import to test files that now use external package
