+++ config_internal_test.go
