+++ entity_binary_sensor_test.go
@@ -1,20 +1,21 @@
-package hal
+package hal_test
 
 import (
 	"testing"
 
+	"github.com/dansimau/hal"
 	"github.com/dansimau/hal/homeassistant"
 	"gotest.tools/v3/assert"
 )
 
 func TestNewBinarySensor(t *testing.T) {
-	sensor := NewBinarySensor("binary_sensor.test")
+	sensor := hal.NewBinarySensor("binary_sensor.test")
 	assert.Equal(t, sensor.GetID(), "binary_sensor.test")
 }
 
 func TestBinarySensor_IsOff(t *testing.T) {
 	t.Run("returns true when state is off", func(t *testing.T) {
-		sensor := NewBinarySensor("binary_sensor.test")
+		sensor := hal.NewBinarySensor("binary_sensor.test")
 		state := homeassistant.State{State: "off"}
 		sensor.SetState(state)
 
@@ -22,7 +23,7 @@ func TestBinarySensor_IsOff(t *testing.T) {
 	})
 
 	t.Run("returns false when state is on", func(t *testing.T) {
-		sensor := NewBinarySensor("binary_sensor.test")
+		sensor := hal.NewBinarySensor("binary_sensor.test")
 		state := homeassistant.State{State: "on"}
 		sensor.SetState(state)
 
@@ -30,7 +31,7 @@ func TestBinarySensor_IsOff(t *testing.T) {
 	})
 
 	t.Run("returns false when state is other value", func(t *testing.T) {
-		sensor := NewBinarySensor("binary_sensor.test")
+		sensor := hal.NewBinarySensor("binary_sensor.test")
 		state := homeassistant.State{State: "unavailable"}
 		sensor.SetState(state)
 
@@ -40,7 +41,7 @@ func TestBinarySensor_IsOff(t *testing.T) {
 
 func TestBinarySensor_IsOn(t *testing.T) {
 	t.Run("returns true when state is on", func(t *testing.T) {
-		sensor := NewBinarySensor("binary_sensor.test")
+		sensor := hal.NewBinarySensor("binary_sensor.test")
 		state := homeassistant.State{State: "on"}
 		sensor.SetState(state)
 
@@ -48,7 +49,7 @@ func TestBinarySensor_IsOn(t *testing.T) {
 	})
 
 	t.Run("returns false when state is off", func(t *testing.T) {
-		sensor := NewBinarySensor("binary_sensor.test")
+		sensor := hal.NewBinarySensor("binary_sensor.test")
 		state := homeassistant.State{State: "off"}
 		sensor.SetState(state)
 
@@ -56,7 +57,7 @@ func TestBinarySensor_IsOn(t *testing.T) {
 	})
 
 	t.Run("returns false when state is other value", func(t *testing.T) {
-		sensor := NewBinarySensor("binary_sensor.test")
+		sensor := hal.NewBinarySensor("binary_sensor.test")
 		state := homeassistant.State{State: "unknown"}
 		sensor.SetState(state)
