+++ entity_input_boolean_test.go
@@ -1,20 +1,21 @@
-package hal
+package hal_test
 
 import (
 	"testing"
 
+	"github.com/dansimau/hal"
 	"github.com/dansimau/hal/homeassistant"
 	"gotest.tools/v3/assert"
 )
 
 func TestNewInputBoolean(t *testing.T) {
-	inputBoolean := NewInputBoolean("input_boolean.test")
+	inputBoolean := hal.NewInputBoolean("input_boolean.test")
 	assert.Equal(t, inputBoolean.GetID(), "input_boolean.test")
 }
 
 func TestInputBoolean_IsOff(t *testing.T) {
 	t.Run("returns true when state is off", func(t *testing.T) {
-		inputBoolean := NewInputBoolean("input_boolean.test")
+		inputBoolean := hal.NewInputBoolean("input_boolean.test")
 		state := homeassistant.State{State: "off"}
 		inputBoolean.SetState(state)
 
@@ -22,7 +23,7 @@ func TestInputBoolean_IsOff(t *testing.T) {
 	})
 
 	t.Run("returns false when state is on", func(t *testing.T) {
-		inputBoolean := NewInputBoolean("input_boolean.test")
+		inputBoolean := hal.NewInputBoolean("input_boolean.test")
 		state := homeassistant.State{State: "on"}
 		inputBoolean.SetState(state)
 
@@ -30,7 +31,7 @@ func TestInputBoolean_IsOff(t *testing.T) {
 	})
 
 	t.Run("returns false when state is other value", func(t *testing.T) {
-		inputBoolean := NewInputBoolean("input_boolean.test")
+		inputBoolean := hal.NewInputBoolean("input_boolean.test")
 		state := homeassistant.State{State: "unavailable"}
 		inputBoolean.SetState(state)
 
@@ -40,7 +41,7 @@ func TestInputBoolean_IsOff(t *testing.T) {
 
 func TestInputBoolean_IsOn(t *testing.T) {
 	t.Run("returns true when state is on", func(t *testing.T) {
-		inputBoolean := NewInputBoolean("input_boolean.test")
+		inputBoolean := hal.NewInputBoolean("input_boolean.test")
 		state := homeassistant.State{State: "on"}
 		inputBoolean.SetState(state)
 
@@ -48,7 +49,7 @@ func TestInputBoolean_IsOn(t *testing.T) {
 	})
 
 	t.Run("returns false when state is off", func(t *testing.T) {
-		inputBoolean := NewInputBoolean("input_boolean.test")
+		inputBoolean := hal.NewInputBoolean("input_boolean.test")
 		state := homeassistant.State{State: "off"}
 		inputBoolean.SetState(state)
 
@@ -56,7 +57,7 @@ func TestInputBoolean_IsOn(t *testing.T) {
 	})
 
 	t.Run("returns false when state is other value", func(t *testing.T) {
-		inputBoolean := NewInputBoolean("input_boolean.test")
+		inputBoolean := hal.NewInputBoolean("input_boolean.test")
 		state := homeassistant.State{State: "unknown"}
 		inputBoolean.SetState(state)
 
@@ -66,22 +67,22 @@ func TestInputBoolean_IsOn(t *testing.T) {
 
 func TestInputBoolean_TurnOn(t *testing.T) {
 	t.Run("returns error when not registered", func(t *testing.T) {
-		inputBoolean := NewInputBoolean("input_boolean.test")
+		inputBoolean := hal.NewInputBoolean("input_boolean.test")
 		err := inputBoolean.TurnOn()
-		assert.Equal(t, err, ErrEntityNotRegistered)
+		assert.Equal(t, err, hal.ErrEntityNotRegistered)
 	})
 
 	t.Run("accepts attributes", func(t *testing.T) {
-		inputBoolean := NewInputBoolean("input_boolean.test")
+		inputBoolean := hal.NewInputBoolean("input_boolean.test")
 		err := inputBoolean.TurnOn(map[string]any{"custom": "value"})
-		assert.Equal(t, err, ErrEntityNotRegistered)
+		assert.Equal(t, err, hal.ErrEntityNotRegistered)
 	})
 }
 
 func TestInputBoolean_TurnOff(t *testing.T) {
 	t.Run("returns error when not registered", func(t *testing.T) {
-		inputBoolean := NewInputBoolean("input_boolean.test")
+		inputBoolean := hal.NewInputBoolean("input_boolean.test")
 		err := inputBoolean.TurnOff()
-		assert.Equal(t, err, ErrEntityNotRegistered)
+		assert.Equal(t, err, hal.ErrEntityNotRegistered)
 	})
 }
