+++ entity_light_test.go
@@ -1,20 +1,21 @@
-package hal
+package hal_test
 
 import (
 	"testing"
 
+	"github.com/dansimau/hal"
 	"github.com/dansimau/hal/homeassistant"
 	"gotest.tools/v3/assert"
 )
 
 func TestNewLight(t *testing.T) {
-	light := NewLight("light.test")
+	light := hal.NewLight("light.test")
 	assert.Equal(t, light.GetID(), "light.test")
 }
 
 func TestLight_GetBrightness(t *testing.T) {
 	t.Run("returns brightness from attributes", func(t *testing.T) {
-		light := NewLight("light.test")
+		light := hal.NewLight("light.test")
 		state := homeassistant.State{
 			Attributes: map[string]any{
 				"brightness": float64(255),
@@ -27,7 +28,7 @@ func TestLight_GetBrightness(t *testing.T) {
 	})
 
 	t.Run("returns 0 when brightness not in attributes", func(t *testing.T) {
-		light := NewLight("light.test")
+		light := hal.NewLight("light.test")
 		state := homeassistant.State{
 			Attributes: map[string]any{},
 		}
@@ -38,7 +39,7 @@ func TestLight_GetBrightness(t *testing.T) {
 	})
 
 	t.Run("returns 0 when brightness is wrong type", func(t *testing.T) {
-		light := NewLight("light.test")
+		light := hal.NewLight("light.test")
 		state := homeassistant.State{
 			Attributes: map[string]any{
 				"brightness": "invalid",
@@ -53,7 +54,7 @@ func TestLight_GetBrightness(t *testing.T) {
 
 func TestLight_IsOn(t *testing.T) {
 	t.Run("returns true when state is on", func(t *testing.T) {
-		light := NewLight("light.test")
+		light := hal.NewLight("light.test")
 		state := homeassistant.State{State: "on"}
 		light.SetState(state)
 
@@ -61,7 +62,7 @@ func TestLight_IsOn(t *testing.T) {
 	})
 
 	t.Run("returns false when state is off", func(t *testing.T) {
-		light := NewLight("light.test")
+		light := hal.NewLight("light.test")
 		state := homeassistant.State{State: "off"}
 		light.SetState(state)
 
@@ -69,7 +70,7 @@ func TestLight_IsOn(t *testing.T) {
 	})
 
 	t.Run("returns false when state is other value", func(t *testing.T) {
-		light := NewLight("light.test")
+		light := hal.NewLight("light.test")
 		state := homeassistant.State{State: "unavailable"}
 		light.SetState(state)
 
@@ -79,40 +80,40 @@ func TestLight_IsOn(t *testing.T) {
 
 func TestLight_TurnOn(t *testing.T) {
 	t.Run("returns error when not registered", func(t *testing.T) {
-		light := NewLight("light.test")
+		light := hal.NewLight("light.test")
 		err := light.TurnOn()
-		assert.Equal(t, err, ErrEntityNotRegistered)
+		assert.Equal(t, err, hal.ErrEntityNotRegistered)
 	})
 
 	t.Run("calls service with attributes", func(t *testing.T) {
 		// This would require mocking the connection and CallService method
 		// For now, we'll test the error case above
-		light := NewLight("light.test")
+		light := hal.NewLight("light.test")
 
 		// Test with attributes
 		err := light.TurnOn(map[string]any{"brightness": 128})
-		assert.Equal(t, err, ErrEntityNotRegistered)
+		assert.Equal(t, err, hal.ErrEntityNotRegistered)
 	})
 }
 
 func TestLight_TurnOff(t *testing.T) {
 	t.Run("returns error when not registered", func(t *testing.T) {
-		light := NewLight("light.test")
+		light := hal.NewLight("light.test")
 		err := light.TurnOff()
-		assert.Equal(t, err, ErrEntityNotRegistered)
+		assert.Equal(t, err, hal.ErrEntityNotRegistered)
 	})
 }
 
 func TestLightGroup_GetID(t *testing.T) {
 	t.Run("returns empty group message for empty group", func(t *testing.T) {
-		lg := LightGroup{}
+		lg := hal.LightGroup{}
 		assert.Equal(t, lg.GetID(), "(empty light group)")
 	})
 
 	t.Run("returns joined IDs for multiple lights", func(t *testing.T) {
-		light1 := NewLight("light.1")
-		light2 := NewLight("light.2")
-		lg := LightGroup{light1, light2}
+		light1 := hal.NewLight("light.1")
+		light2 := hal.NewLight("light.2")
+		lg := hal.LightGroup{light1, light2}
 
 		assert.Equal(t, lg.GetID(), "light.1, light.2")
 	})
@@ -120,13 +121,13 @@ func TestLightGroup_GetID(t *testing.T) {
 
 func TestLightGroup_GetBrightness(t *testing.T) {
 	t.Run("returns 0 for empty group", func(t *testing.T) {
-		lg := LightGroup{}
+		lg := hal.LightGroup{}
 		assert.Equal(t, lg.GetBrightness(), float64(0))
 	})
 
 	t.Run("returns brightness of first light", func(t *testing.T) {
-		light1 := NewLight("light.1")
-		light2 := NewLight("light.2")
+		light1 := hal.NewLight("light.1")
+		light2 := hal.NewLight("light.2")
 
 		state1 := homeassistant.State{Attributes: map[string]any{"brightness": float64(100)}}
 		state2 := homeassistant.State{Attributes: map[string]any{"brightness": float64(200)}}
@@ -134,24 +135,24 @@ func TestLightGroup_GetBrightness(t *testing.T) {
 		light1.SetState(state1)
 		light2.SetState(state2)
 
-		lg := LightGroup{light1, light2}
+		lg := hal.LightGroup{light1, light2}
 		assert.Equal(t, lg.GetBrightness(), float64(100))
 	})
 }
 
 func TestLightGroup_GetState(t *testing.T) {
 	t.Run("returns empty state for empty group", func(t *testing.T) {
-		lg := LightGroup{}
+		lg := hal.LightGroup{}
 		state := lg.GetState()
 		assert.DeepEqual(t, state, homeassistant.State{})
 	})
 
 	t.Run("returns state of first light", func(t *testing.T) {
-		light1 := NewLight("light.1")
+		light1 := hal.NewLight("light.1")
 		expectedState := homeassistant.State{State: "on", Attributes: map[string]any{"brightness": float64(255)}}
 		light1.SetState(expectedState)
 
-		lg := LightGroup{light1}
+		lg := hal.LightGroup{light1}
 		state := lg.GetState()
 		assert.DeepEqual(t, state, expectedState)
 	})
@@ -159,9 +160,9 @@ func TestLightGroup_GetState(t *testing.T) {
 
 func TestLightGroup_SetState(t *testing.T) {
 	t.Run("sets state on all lights in group", func(t *testing.T) {
-		light1 := NewLight("light.1")
-		light2 := NewLight("light.2")
-		lg := LightGroup{light1, light2}
+		light1 := hal.NewLight("light.1")
+		light2 := hal.NewLight("light.2")
+		lg := hal.LightGroup{light1, light2}
 
 		newState := homeassistant.State{State: "on", Attributes: map[string]any{"brightness": float64(128)}}
 		lg.SetState(newState)
@@ -173,29 +174,29 @@ func TestLightGroup_SetState(t *testing.T) {
 
 func TestLightGroup_IsOn(t *testing.T) {
 	t.Run("returns true when all lights are on", func(t *testing.T) {
-		light1 := NewLight("light.1")
-		light2 := NewLight("light.2")
+		light1 := hal.NewLight("light.1")
+		light2 := hal.NewLight("light.2")
 
 		light1.SetState(homeassistant.State{State: "on"})
 		light2.SetState(homeassistant.State{State: "on"})
 
-		lg := LightGroup{light1, light2}
+		lg := hal.LightGroup{light1, light2}
 		assert.Equal(t, lg.IsOn(), true)
 	})
 
 	t.Run("returns false when any light is off", func(t *testing.T) {
-		light1 := NewLight("light.1")
-		light2 := NewLight("light.2")
+		light1 := hal.NewLight("light.1")
+		light2 := hal.NewLight("light.2")
 
 		light1.SetState(homeassistant.State{State: "on"})
 		light2.SetState(homeassistant.State{State: "off"})
 
-		lg := LightGroup{light1, light2}
+		lg := hal.LightGroup{light1, light2}
 		assert.Equal(t, lg.IsOn(), false)
 	})
 
 	t.Run("returns true for empty group", func(t *testing.T) {
-		lg := LightGroup{}
+		lg := hal.LightGroup{}
 		assert.Equal(t, lg.IsOn(), true)
 	})
 }
@@ -203,17 +204,17 @@ func TestLightGroup_IsOn(t *testing.T) {
 func TestLightGroup_TurnOn(t *testing.T) {
 	t.Run("returns nil when no errors", func(t *testing.T) {
 		// Empty group should not error
-		lg := LightGroup{}
+		lg := hal.LightGroup{}
 		err := lg.TurnOn()
 		assert.NilError(t, err)
 	})
 
 	t.Run("collects errors from individual lights", func(t *testing.T) {
-		light1 := NewLight("light.1")
-		light2 := NewLight("light.2")
-		lg := LightGroup{light1, light2}
+		light1 := hal.NewLight("light.1")
+		light2 := hal.NewLight("light.2")
+		lg := hal.LightGroup{light1, light2}
 
-		// Both lights should return ErrEntityNotRegistered
+		// Both lights should return hal.ErrEntityNotRegistered
 		err := lg.TurnOn()
 		// With 2 errors, should return a joined error
 		assert.ErrorContains(t, err, "entity not registered")
@@ -222,15 +223,15 @@ func TestLightGroup_TurnOn(t *testing.T) {
 
 func TestLightGroup_TurnOff(t *testing.T) {
 	t.Run("returns nil when no errors", func(t *testing.T) {
-		lg := LightGroup{}
+		lg := hal.LightGroup{}
 		err := lg.TurnOff()
 		assert.NilError(t, err)
 	})
 
 	t.Run("collects errors from individual lights", func(t *testing.T) {
-		light1 := NewLight("light.1")
-		light2 := NewLight("light.2")
-		lg := LightGroup{light1, light2}
+		light1 := hal.NewLight("light.1")
+		light2 := hal.NewLight("light.2")
+		lg := hal.LightGroup{light1, light2}
 
 		err := lg.TurnOff()
 		// With 2 errors, should return joined error
@@ -240,12 +241,12 @@ func TestLightGroup_TurnOff(t *testing.T) {
 
 func TestLightGroup_BindConnection(t *testing.T) {
 	t.Run("binds connection to all lights", func(t *testing.T) {
-		light1 := NewLight("light.1")
-		light2 := NewLight("light.2")
-		lg := LightGroup{light1, light2}
+		light1 := hal.NewLight("light.1")
+		light2 := hal.NewLight("light.2")
+		lg := hal.LightGroup{light1, light2}
 
 		// Create a mock connection (would need proper setup in real scenario)
-		conn := &Connection{}
+		conn := &hal.Connection{}
 		lg.BindConnection(conn)
 
 		// Verify connection was bound (in real test, would check that connection property was set)
